pyproject.toml
[project]
name = "translation-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.2",
    "fastapi[standard]>=0.116.0",
    "jinja2>=3.1.6",
    "openai>=1.93.1",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "sqlmodel>=0.0.24",
]
